style: continue making footer style
