+++ package.json
@@ -17,6 +17,7 @@
     "@types/react-dom": "^19.0.3",
     "animate.css": "^4.1.1",
     "cra-template-typescript": "1.2.0",
+    "framer-motion": "^12.0.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-router-dom": "^6.26.1",
